+++ platformio/commands/list.py
@@ -3,14 +3,19 @@
 
 from click import command, echo, style
 
-from platformio.pkgmanager import PackageManager
+from platformio.platforms.base import PlatformFactory
 
 
 @command("list", short_help="List installed platforms")
 def cli():
 
-    for name, pkgs in PackageManager.get_installed().items():
+    installed_platforms = PlatformFactory.get_platforms(
+        installed=True).keys()
+    sorted(installed_platforms)
+
+    for platform in installed_platforms:
+        p = PlatformFactory().newPlatform(platform)
         echo("{name:<20} with packages: {pkgs}".format(
-            name=style(name, fg="cyan"),
-            pkgs=", ".join(pkgs.keys())
+            name=style(p.get_name(), fg="cyan"),
+            pkgs=", ".join(p.get_installed_packages())
         ))