+++ platformio/platforms/timsp430.py
@@ -1,8 +1,6 @@
 # Copyright (C) Ivan Kravets <me@ikravets.com>
 # See LICENSE for details.
 
-from os.path import join
-
 from platformio.platforms.base import BasePlatform
 
 
@@ -15,19 +13,17 @@ class Timsp430Platform(BasePlatform):
     PACKAGES = {
 
         "toolchain-timsp430": {
-            "path": join("tools", "toolchain"),
             "alias": "toolchain",
             "default": True
         },
 
         "tool-mspdebug": {
-            "path": join("tools", "mspdebug"),
             "alias": "uploader",
             "default": True
         },
 
         "framework-energiamsp430": {
-            "path": join("frameworks", "energia"),
+            "alias": "framework",
             "default": True
         }
     }